+++ src/__tests__/cli-e2e-docs-config.test.ts
@@ -312,8 +312,8 @@ describe("CLI docsConfig tests", () => {
             expect(1).toEqual(2);
           } catch (e) {
             expect((e as Error).message).toContain("exited with non-zero code: 1");
-            expect((e as spawnAsync.SpawnResult).output[0]).toContain(
-              "Example ./src/__tests__/tmpTestData-cli-e2e-docs-config/examples5/example1.json is not valid",
+            expect((e as spawnAsync.SpawnResult).stderr).toContain(
+              "Example src/__tests__/tmpTestData-cli-e2e-docs-config/examples5/example1.json is not valid",
             );
           }
         });